Move CLI to seperate file in src directory
- Move CLI files from server directory to its own in source.
